(blind) fix for gcc warning on some platforms
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@25419 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ src/common/intl.cpp
@@ -2423,14 +2423,14 @@ const wxChar *wxLocale::GetString(const wxChar *szOrigString,
             {
                 wxLogTrace(_T("i18n"),
                            _T("string '%s'[%d] not found in domain '%s' for locale '%s'."),
-                           szOrigString, n, szDomain, m_strLocale.c_str());
+                           szOrigString, (int)n, szDomain, m_strLocale.c_str());
 
             }
             else
             {
                 wxLogTrace(_T("i18n"),
                            _T("string '%s'[%d] not found in locale '%s'."),
-                           szOrigString, n, m_strLocale.c_str());
+                           szOrigString, (int)n, m_strLocale.c_str());
             }
         }
 #endif // __WXDEBUG__
